minor: 转移主机至待回收模块调整
